requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
Pillow>=8.3.0
seaborn>=0.11.0